requirements.txt
Django==4.2.7
mysqlclient==2.2.0
google-generativeai==0.3.0
python-dotenv==1.0.0
Pillow==10.1.0
pillow-heif==0.13.0

